DJANGO/requirements.txt
flask
requests
Django
scapy
gunicorn
djangorestframework